refactor gbq tests
