+++ tests/unit/conftest.py
@@ -9,6 +9,36 @@ from unittest import mock
 import pytest
 
 
+def mock_get_credentials(*args, **kwargs):
+    import google.auth.credentials
+
+    mock_credentials = mock.create_autospec(google.auth.credentials.Credentials)
+    return mock_credentials, "default-project"
+
+
+@pytest.fixture
+def mock_service_account_credentials():
+    import google.oauth2.service_account
+
+    mock_credentials = mock.create_autospec(google.oauth2.service_account.Credentials)
+    return mock_credentials
+
+
+@pytest.fixture
+def mock_compute_engine_credentials():
+    import google.auth.compute_engine
+
+    mock_credentials = mock.create_autospec(google.auth.compute_engine.Credentials)
+    return mock_credentials
+
+
+@pytest.fixture(autouse=True)
+def no_auth(monkeypatch):
+    import pydata_google_auth
+
+    monkeypatch.setattr(pydata_google_auth, "default", mock_get_credentials)
+
+
 @pytest.fixture(autouse=True, scope="function")
 def reset_context():
     import pandas_gbq
