+++ tests/unit/test_auth.py
@@ -47,16 +47,12 @@ def test_get_credentials_load_user_no_default(monkeypatch):
     import google.auth.credentials
     import pydata_google_auth.cache
 
-    def mock_default_credentials(scopes=None, request=None):
-        return (None, None)
-
-    monkeypatch.setattr(google.auth, "default", mock_default_credentials)
     mock_user_credentials = mock.create_autospec(google.auth.credentials.Credentials)
 
-    mock_cache = mock.create_autospec(pydata_google_auth.cache.CredentialsCache)
-    mock_cache.load.return_value = mock_user_credentials
+    def mock_default_credentials(scopes, **kwargs):
+        return (mock_user_credentials, None)
 
-    monkeypatch.setattr(auth, "get_credentials_cache", lambda _: mock_cache)
+    monkeypatch.setattr(pydata_google_auth, "default", mock_default_credentials)
 
     credentials, project = auth.get_credentials()
     assert project is None
